bug: merged folders into renamed provider
