--- src/context/EarnManagerProvider/protocols/yearn/constants/vaults.ts
@@ -1,17 +1,0 @@
-export type YearnVault = {
-  contractAddress: string
-  routerAddress: string
-  decimals: number
-  symbol: string
-  depositToken: string
-}
-
-export const vaults: YearnVault[] = [
-  {
-    contractAddress: '0x5f18C75AbDAe578b483E5F43f12a39cF75b973a9',
-    routerAddress: '',
-    decimals: 6,
-    symbol: 'yUSDC',
-    depositToken: '0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48'
-  }
-]
